+++ src/main/java/es/in2/walletdata/configuration/CorsGlobalConfig.java
@@ -7,8 +7,8 @@ import org.springframework.web.reactive.config.CorsRegistry;
 import org.springframework.web.reactive.config.EnableWebFlux;
 import org.springframework.web.reactive.config.WebFluxConfigurer;
 
-import static es.in2.walletdata.utils.ApiUtils.ALLOWED_METHODS;
-import static es.in2.walletdata.utils.ApiUtils.GLOBAL_ENDPOINTS_API;
+import static es.in2.walletdata.utils.Utils.ALLOWED_METHODS;
+import static es.in2.walletdata.utils.Utils.GLOBAL_ENDPOINTS_API;
 
 @Configuration
 @EnableWebFlux
@@ -19,7 +19,7 @@ public class CorsGlobalConfig implements WebFluxConfigurer {
     @Override
     public void addCorsMappings(CorsRegistry corsRegistry) {
         corsRegistry.addMapping(GLOBAL_ENDPOINTS_API)
-                .allowedOrigins(walletDrivingApplicationProperties.url())
+                .allowedOrigins(walletDrivingApplicationProperties.domain())
                 .allowedMethods(ALLOWED_METHODS)
                 .maxAge(3600);
     }