+++ src/main/java/es/in2/walletdata/controller/UserController.java
@@ -3,7 +3,6 @@ package es.in2.walletdata.controller;
 import es.in2.walletdata.domain.UserAttribute;
 import es.in2.walletdata.domain.UserRequest;
 import es.in2.walletdata.facade.UserDataFacadeService;
-import es.in2.walletdata.utils.ApplicationUtils;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.validation.Valid;
@@ -13,6 +12,8 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
+
+import static es.in2.walletdata.utils.Utils.getUserIdFromToken;
 
 @RestController
 @RequestMapping("/api/v2/users")
@@ -51,7 +52,7 @@ public class UserController {
     public Mono<UserAttribute> getUserDataByUserId(@RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader) {
         log.debug("UserController.getUserDataByUserId");
         log.debug("VerifiableCredentialController.getSelectableVCs()");
-        return ApplicationUtils.getUserIdFromToken(authorizationHeader)
+        return getUserIdFromToken(authorizationHeader)
                 .flatMap(userDataFacadeService::getUserDataByUserId);
     }